Adjust for the elimination of the "float" and "complex" types.
git-svn-id: https://swig.svn.sourceforge.net/svnroot/swig/trunk@12487 626c5289-ae23-0410-ae9c-e8d60b6d4f22

+++ Examples/test-suite/go/overload_extend_runme.go
@@ -16,7 +16,7 @@ func main() {
 	if f.Test(float64(3), float64(2)).(float64) != 5 {
 		panic(0)
 	}
-	if f.Test(float64(3.0)).(float64) != 1003 {
+	if f.Test(3.0).(float64) != 1003 {
 		panic(0)
 	}
 }
